pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shopping-assistant"
version = "0.1.0"
description = "Conversational CLI shopping assistant that clarifies needs and researches real products."
authors = [
  { name = "Shopping Assistant" }
]
requires-python = ">=3.10"
dependencies = [
  "openai>=1.30.0",
  "python-dotenv>=1.0.1",
  "rich>=13.7.1",
  "typer>=0.12.3",
  "tavily-python>=0.3.5",
  "httpx>=0.27.0",
  "pydantic>=2.7.1",
  "pydantic-settings>=2.2.1"
]

[project.scripts]
shopping-assistant = "shopping_assistant.cli:app"
